+++ Filter/Exception/StoreDoesNotExistException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Hackathon\EAVCleaner\Filter\Exception;
+
+use Symfony\Component\Console\Exception\InvalidOptionException;
+
+class StoreDoesNotExistException extends InvalidOptionException
+{
+
+}